refactor: forgot some stories folders and modify build for new folder structure
