+++ package.json
@@ -109,44 +109,45 @@
     "styled-components": "^5.x"
   },
   "scripts": {
-    "dev": "concurrently 'yarn storybook' 'yarn proxy'",
-    "dev:open": "concurrently 'yarn storybook' 'yarn proxy' 'yarn open'",
-    "prebuild:theme": "echo 'start generating styles'",
+    "build:ci": "yarn build:theme && ts-node build-utils/build.ts && yarn build:dts",
+    "build:colors": "ts-node build-utils/css/buildColorConstants.ts",
+    "build:dts": "ttsc -p ./tsconfig.emitDeclarationOnly.json",
+    "build:logic-constant": "ts-node build-utils/css/buildLogicConstants.ts",
+    "build:prep:publish": "ts-node build-utils/setupPackage.ts",
     "build:theme": "ts-node build-utils/css/buildTheme.ts",
-    "postbuild:theme": "echo 'done generating styles'",
-    "prebuild": "echo 'start cleaning up' && rm -rf dist && echo 'done cleaning up'",
     "build": "yarn && yarn ci:concurrent && yarn build:theme && ts-node build-utils/build.ts && yarn build:dts",
+    "ci:concurrent": "concurrently 'yarn tsc' 'yarn lint:prod' 'yarn test:prod' --group --raw --success='all'",
+    "ci": "yarn tsc && yarn lint:prod && yarn test:prod",
+    "dev:open": "concurrently 'yarn storybook' 'yarn proxy' 'yarn open'",
+    "dev": "concurrently 'yarn storybook' 'yarn proxy'",
+    "lint:prod": "eslint src --ext .ts,.tsx -c ./.eslintrc.prod.cjs",
+    "lint": "eslint src --ext .ts,.tsx",
+    "open": "open https://amino-dev.zonos.com:6007",
+    "postbuild:ci": "yarn postbuild",
+    "postbuild:dts": "echo 'done build dts'",
+    "postbuild:prep:publish": "echo 'done building for publishing'",
+    "postbuild:theme": "echo 'done generating styles'",
     "postbuild": "cp -r ./src/styles/*.css ./dist && yarn build:prep:publish",
     "prebuild:ci": "yarn prebuild",
-    "build:ci": "yarn build:theme && ts-node build-utils/build.ts && yarn build:dts",
-    "postbuild:ci": "yarn postbuild",
     "prebuild:dts": "echo 'building dts'",
-    "build:dts": "ttsc -p ./tsconfig.emitDeclarationOnly.json",
-    "postbuild:dts": "echo 'done build dts'",
     "prebuild:prep:publish": "echo 'building for publishing'",
-    "build:prep:publish": "ts-node build-utils/setupPackage.ts",
-    "postbuild:prep:publish": "echo 'done building for publishing'",
-    "template:logic-constant": "ts-node build-utils/css/generateLogicConstTemplate.ts",
-    "build:logic-constant": "ts-node build-utils/css/buildLogicConstants.ts",
-    "build:colors": "ts-node build-utils/css/buildColorConstants.ts",
-    "test": "jest --watchAll --testPathIgnorePatterns storyshots",
-    "test:prod": "jest --testPathIgnorePatterns storyshots",
-    "test:visual": "jest ./.storybook/storyshots",
-    "lint": "eslint src --ext .ts,.tsx",
-    "lint:prod": "eslint src --ext .ts,.tsx -c ./.eslintrc.prod.cjs",
-    "ci": "yarn tsc && yarn lint:prod && yarn test:prod",
-    "ci:concurrent": "concurrently 'yarn tsc' 'yarn lint:prod' 'yarn test:prod' --group --raw --success='all'",
-    "open": "open https://amino-dev.zonos.com:6007",
-    "proxy": "local-ssl-proxy --source 6007 --target 6006",
-    "svgs:format": "eslint --fix svgReact --ext .tsx -c ./.eslintrc.prod.cjs",
+    "prebuild:theme": "echo 'start generating styles'",
+    "prebuild": "echo 'start cleaning up' && rm -rf dist && echo 'done cleaning up'",
     "presvgs:react:flags": "ts-node ./svgReact/flags/runGenerateSvgs.ts",
-    "svgs:react:flags": "ts-node ./svgReact/flags/runCreateComponentsFromSvgs.ts",
-    "svgs:react:icons": "ts-node ./svgReact/icons",
+    "proxy": "local-ssl-proxy --source 6007 --target 6006",
+    "push:tags": "git push --follow-tags",
     "start": "yarn build-storybook && yarn storybook",
     "storybook": "start-storybook -p 6006 --no-open --quiet",
-    "version:patch": "yarn build && npm version patch -m 'Bump version %s' && yarn push:tags",
-    "version:minor": "yarn build && npm version minor -m 'Bump version %s' && yarn push:tags",
+    "svgs:format": "eslint --fix svgReact --ext .tsx -c ./.eslintrc.prod.cjs",
+    "svgs:react:flags": "ts-node ./svgReact/flags/runCreateComponentsFromSvgs.ts",
+    "svgs:react:icons": "ts-node ./svgReact/icons",
+    "template:logic-constant": "ts-node build-utils/css/generateLogicConstTemplate.ts",
+    "test:prod": "jest --testPathIgnorePatterns storyshots",
+    "test:visual": "jest ./.storybook/storyshots",
+    "test": "jest --watchAll --testPathIgnorePatterns storyshots",
+    "tsc": "tsc --noEmit",
     "version:major": "yarn build && npm version major -m 'Bump version %s' && yarn push:tags",
-    "push:tags": "git push --follow-tags"
+    "version:minor": "yarn build && npm version minor -m 'Bump version %s' && yarn push:tags",
+    "version:patch": "yarn build && npm version patch -m 'Bump version %s' && yarn push:tags"
   }
 }
