+++ src/all.ts
@@ -635,7 +635,6 @@ import './components/menu/MenuItem';
 import './components/notice/Notice';
 import './components/radio/Radio';
 import './components/radio/RadioGroup';
-import './components/slider/Slider';
 import './components/rest-state/RestState';
 import './components/rich-checkbox/RichCheckbox';
 import './components/rich-radio/RichRadio';
@@ -648,6 +647,7 @@ import './components/select/Select';
 import './components/select/CountrySelect';
 import './components/skeleton/Skeleton';
 import './components/slide-over/SlideOver';
+import './components/slider/Slider';
 import './components/sortable-list/SortableList';
 import './components/sortable-list/SortableListDeps';
 import './components/sortable-list/SortableListItem';
