+++ src/components/rich-checkbox/__stories__/RichCheckbox.stories.tsx
