+++ src/components/rounded-icon/__stories__/RoundedIcon.stories.tsx
