+++ src/components/select/__stories__/CountryMultiSelect.stories.tsx
@@ -7,8 +7,8 @@ import {
 } from 'src/components/select/CountryMultiSelect';
 import styled from 'styled-components';
 
-import { getCountryUrls } from './getCountryUrls';
-import { useCountryOptions } from './useCountryOptions';
+import { getCountryUrls } from './getCountryUrls.stories';
+import { useCountryOptions } from './useCountryOptions.stories';
 
 const StyledWrapper = styled.div`
   width: 412px;
